requirements.txt
streamlit
pymupdf
mistralai
python-dotenv
